feat: added report output and other options
This adds a new report system, similar to what Photorec outputs during a
carve. An option has been added to the CLI to disable file extraction in
favor of only generating the report. Both this report and the audit
report can be suppressed with other options. Non-WASM-friendly crates
are only included in the CLI so the file carving can eventually be set
up to work in a browser environment via WASM.